Refactor to f32.

+++ src/camera.rs
@@ -12,7 +12,7 @@ pub struct Camera {
     u: Vec3,
     v: Vec3,
     w: Vec3,
-    lens_radius: f64,
+    lens_radius: f32,
 }
 
 impl Camera {
@@ -20,10 +20,10 @@ impl Camera {
         lookfrom: Point3,
         lookat: Point3,
         vup: Vec3,
-        vfov: f64,
-        aspect_ratio: f64,
-        aperature: f64,
-        focus_dist: f64,
+        vfov: f32,
+        aspect_ratio: f32,
+        aperature: f32,
+        focus_dist: f32,
     ) -> Camera {
         let theta = degrees_to_radians(vfov);
         let h = (theta / 2.0).tan();
@@ -52,7 +52,7 @@ impl Camera {
         }
     }
 
-    pub fn get_ray(&self, s: f64, t: f64, rng: &mut Rand) -> Ray {
+    pub fn get_ray(&self, s: f32, t: f32, rng: &mut Rand) -> Ray {
         let rd = self.lens_radius * Vec3::random_in_unit_disk(rng);
         let offset = self.u * rd.x() + self.v * rd.y();
